mcp-server/apigee-mcp-server/apigee-api-to-mcp/requirements.txt
fastapi
fastmcp
PyYAML>=6.0.2
requests>=2.32.4
uvicorn>=0.34.0